requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
faiss-cpu
langchain
langchain-community
langchain-text-splitters
sentence-transformers
pypdf
